Add BatteryBank for max joltage calculation
Introduce the `BatteryBank` struct to encapsulate battery-related logic
and implement a `max_joltage` method for calculating the maximum joltage
from a given number of digits. Refactor `part01` to use this new
structure and implement `part02` to calculate total joltage for 12
digits. Update the test for `part02` with the expected result.

This works, but it's essentially bruteforcing the answer, and for real
data part02 takes way too long.